+++ CHANGELOG.md
@@ -48,3 +48,4 @@
 
 ### 0.5.1
 * auto-generate autocomplete test descriptions
+* properly kill server process after tests